Добавляет к проекту webpack, webpack-dev-server. Настраивает его. Добавляет файлы сборки в gitignore. Изменяет настройки eslint. Разбивает проект на прото-модули
