+++ src/components/board/board.js
@@ -1,0 +1,5 @@
+const createBoardTemplate = () => {
+  return `<section class="board container"></section>`;
+};
+
+export {createBoardTemplate};
