+++ src/components/board/more-button.js
@@ -1,0 +1,5 @@
+const createBoardMoreButtonTemplate = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
+
+export {createBoardMoreButtonTemplate};
