+++ src/components/board/task.js
@@ -1,0 +1,6 @@
+const createBoardTaskTemplate = () => {
+  return `<div class="board__tasks">
+  </div>`;
+};
+
+export {createBoardTaskTemplate};
